refactor(http): unify-http-clients-for-opamp-and-oauth
