+++ agent-control/src/agent_control/config_storer/store.rs
@@ -99,7 +99,7 @@ where
                 FileFormat::Yaml,
             ))
             // Add in settings from the environment (with a prefix of `NR_` and separator double underscore, `__`)
-            // Eg.. `NR_LOG__USE_DEBUG=1 ./target/app` would set the `log.use_debug` key
+            // Eg.. `NR_LOG_USE_DEBUG=1 ./target/app` would set the `log.use_debug` key
             // We use double underscore because we already use snake_case for the config keys.
             .add_source(
                 Environment::with_prefix(AGENT_CONTROL_CONFIG_ENV_VAR_PREFIX)
