+++ agent-control/src/http.rs
@@ -1,4 +1,3 @@
+pub mod client;
 pub mod config;
-pub mod proxy;
-pub mod reqwest;
 pub mod tls;
