+++ agent-control/src/opamp/auth.rs
@@ -1,3 +1,2 @@
 pub mod config;
-mod http_client;
 pub mod token_retriever;
